Modify to reuse PikachuVolleyballReplay object ..
Before this modification, when we scrubbed quite a lot the replay,
the audio began to emit crackling sounds when I tested on my smartphone.
I don't know the exact reason but it surely had something to do with
the fact that the PikaAudio object created too much on the same sound
files when doing scrub.