Remove unnecessary this.willMoveScrubber
this.willMoveScrubber can be safely removed.
It was first intoduced by a worry of calling moveScrubberTo too
much, so that updating scrubberRangeInput.value and letting the browser
render the updated scrubber too much often, in seekFrame which calls
it by calling gameLoopSilent multiple times. However, this worry is
not legitimate because

"Rendering never happens while the engine executes a task. It doesn’t
matter if the task takes a long time. Changes to the DOM are painted
only after the task is complete."
Reference: https://javascript.info/event-loop

So scrubberRangeInput.value is rendered only after the engine executes
the designated task. So while the engine executes, the
scrubberRangeInput.value can be assigned multiple times with different
values without affecting the rendering.