Enhance network test to detect no host candidate
and make notice message boxes for it.

I found the following things while troubeshooting with "안녕하세요":
- Some browser extensions can block WebRTC
- And some VPN services also may block WebRTC
And for these cases, the host candidate can be not gathered.
Some event corresponding to the cases above can block only server
reflexive candidate and does not block host candidate? I don't know.
But for now, I show messages for the two cases just when the browser
does not gather both the server reflexive candidate and the host
candidate.